refactor(web): refactors element-attachment code as own module
